Reorganize test code
