+++ src/world/World.test.ts
@@ -1,4 +1,3 @@
-import type { Shape } from '../shapes/Shape'
 import { Vec3 } from '../math/Vec3'
 import { Box } from '../shapes/Box'
 import { Sphere } from '../shapes/Sphere'
@@ -7,7 +6,7 @@ import { World } from '../world/World'
 import { ArrayCollisionMatrix } from '../collision/ArrayCollisionMatrix'
 import { ObjectCollisionMatrix } from '../collision/ObjectCollisionMatrix'
 import { RaycastResult } from '../collision/RaycastResult'
-import { testCollisionMatrix } from '../../test/helpers'
+import { testCollisionMatrix } from './testHelpers'
 
 describe('World', () => {
   test('clearForces', () => {
